+++ libs/middletier/MockMiddleTierAdminTest.cpp
@@ -29,7 +29,7 @@ TEST_F(cta_client_MockMiddleTierAdminTest, createStorageClass_new) {
   }
 
   const std::string name = "TestStorageClass";
-  const uint8_t nbCopies = 2;
+  const uint16_t nbCopies = 2;
   const std::string comment = "Comment";
   ASSERT_NO_THROW(adminApi.createStorageClass(requester, name, nbCopies,
     comment));
@@ -61,7 +61,7 @@ TEST_F(cta_client_MockMiddleTierAdminTest,
   }
 
   const std::string name = "TestStorageClass";
-  const uint8_t nbCopies = 2;
+  const uint16_t nbCopies = 2;
   const std::string comment = "Comment";
   ASSERT_NO_THROW(adminApi.createStorageClass(requester, name, nbCopies, comment));
 
@@ -128,7 +128,7 @@ TEST_F(cta_client_MockMiddleTierAdminTest, deleteStorageClass_existing) {
   }
 
   const std::string name = "TestStorageClass";
-  const uint8_t nbCopies = 2;
+  const uint16_t nbCopies = 2;
   const std::string comment = "Comment";
   ASSERT_NO_THROW(adminApi.createStorageClass(requester, name, nbCopies, comment));
 
@@ -167,7 +167,7 @@ TEST_F(cta_client_MockMiddleTierAdminTest,
   }
 
   const std::string storageClassName = "TestStorageClass";
-  const uint8_t nbCopies = 2;
+  const uint16_t nbCopies = 2;
   const std::string comment = "Comment";
   ASSERT_NO_THROW(adminApi.createStorageClass(requester, storageClassName, nbCopies,
     comment));
@@ -226,7 +226,7 @@ TEST_F(cta_client_MockMiddleTierAdminTest, deleteStorageClass_in_use_by_route) {
   }
 
   const std::string storageClassName = "TestStorageClass";
-  const uint8_t nbCopies = 2;
+  const uint16_t nbCopies = 2;
   const std::string comment = "Comment";
   ASSERT_NO_THROW(adminApi.createStorageClass(requester, storageClassName, nbCopies,
     comment));
@@ -258,7 +258,7 @@ TEST_F(cta_client_MockMiddleTierAdminTest, deleteStorageClass_in_use_by_route) {
     ASSERT_EQ(tapePoolName, tapePool.getName());
   }
 
-  const uint8_t copyNb = 1;
+  const uint16_t copyNb = 1;
   ASSERT_NO_THROW(adminApi.createArchiveRoute(requester, storageClassName,
     copyNb, tapePoolName, comment));
 
@@ -345,7 +345,7 @@ TEST_F(cta_client_MockMiddleTierAdminTest, deleteTapePool_in_use) {
   const std::string storageClassName = "TestStorageClass";
   const std::string comment = "Comment";
   {
-    const uint8_t nbCopies = 2;
+    const uint16_t nbCopies = 2;
     ASSERT_NO_THROW(adminApi.createStorageClass(requester, storageClassName,
       nbCopies, comment));
   }
@@ -356,7 +356,7 @@ TEST_F(cta_client_MockMiddleTierAdminTest, deleteTapePool_in_use) {
   ASSERT_NO_THROW(adminApi.createTapePool(requester, tapePoolName, nbDrives,
     nbPartialTapes, comment));
 
-  const uint8_t copyNb = 1;
+  const uint16_t copyNb = 1;
   ASSERT_NO_THROW(adminApi.createArchiveRoute(requester, storageClassName,
     copyNb, tapePoolName, comment));
 
@@ -403,7 +403,7 @@ TEST_F(cta_client_MockMiddleTierAdminTest, createArchiveRoute_new) {
   const std::string storageClassName = "TestStorageClass";
   const std::string comment = "Comment";
   {
-    const uint8_t nbCopies = 2;
+    const uint16_t nbCopies = 2;
     ASSERT_NO_THROW(adminApi.createStorageClass(requester, storageClassName,
       nbCopies, comment));
   }
@@ -414,7 +414,7 @@ TEST_F(cta_client_MockMiddleTierAdminTest, createArchiveRoute_new) {
   ASSERT_NO_THROW(adminApi.createTapePool(requester, tapePoolName, nbDrives,
     nbPartialTapes, comment));
 
-  const uint8_t copyNb = 1;
+  const uint16_t copyNb = 1;
   ASSERT_NO_THROW(adminApi.createArchiveRoute(requester, storageClassName,
     copyNb, tapePoolName, comment));
 
@@ -448,7 +448,7 @@ TEST_F(cta_client_MockMiddleTierAdminTest,
   const std::string storageClassName = "TestStorageClass";
   const std::string comment = "Comment";
   {
-    const uint8_t nbCopies = 2;
+    const uint16_t nbCopies = 2;
     ASSERT_NO_THROW(adminApi.createStorageClass(requester, storageClassName,
       nbCopies, comment));
   }
@@ -459,7 +459,7 @@ TEST_F(cta_client_MockMiddleTierAdminTest,
   ASSERT_NO_THROW(adminApi.createTapePool(requester, tapePoolName, nbDrives,
     nbPartialTapes, comment));
 
-  const uint8_t copyNb = 1;
+  const uint16_t copyNb = 1;
   ASSERT_NO_THROW(adminApi.createArchiveRoute(requester, storageClassName,
     copyNb, tapePoolName, comment));
 
@@ -495,7 +495,7 @@ TEST_F(cta_client_MockMiddleTierAdminTest, deleteArchiveRoute_existing) {
   const std::string storageClassName = "TestStorageClass";
   const std::string comment = "Comment";
   {
-    const uint8_t nbCopies = 2;
+    const uint16_t nbCopies = 2;
     ASSERT_NO_THROW(adminApi.createStorageClass(requester, storageClassName,
       nbCopies, comment));
   }
@@ -506,7 +506,7 @@ TEST_F(cta_client_MockMiddleTierAdminTest, deleteArchiveRoute_existing) {
   ASSERT_NO_THROW(adminApi.createTapePool(requester, tapePoolName, nbDrives,
     nbPartialTapes, comment));
 
-  const uint8_t copyNb = 1;
+  const uint16_t copyNb = 1;
   ASSERT_NO_THROW(adminApi.createArchiveRoute(requester, storageClassName,
     copyNb, tapePoolName, comment));
 
@@ -548,7 +548,7 @@ TEST_F(cta_client_MockMiddleTierAdminTest, deleteArchiveRoute_non_existing) {
   const std::string storageClassName = "TestStorageClass";
   const std::string comment = "Comment";
   {
-    const uint8_t nbCopies = 2;
+    const uint16_t nbCopies = 2;
     ASSERT_NO_THROW(adminApi.createStorageClass(requester, storageClassName,
       nbCopies, comment));
   }
@@ -559,7 +559,7 @@ TEST_F(cta_client_MockMiddleTierAdminTest, deleteArchiveRoute_non_existing) {
   ASSERT_NO_THROW(adminApi.createTapePool(requester, tapePoolName, nbDrives,
     nbPartialTapes, comment));
 
-  const uint8_t copyNb = 1;
+  const uint16_t copyNb = 1;
   ASSERT_THROW(adminApi.deleteArchiveRoute(requester, tapePoolName, copyNb),
     std::exception);
 }
