+++ libs/middletier/SqliteDatabase.hpp
@@ -32,9 +32,9 @@ public:
   
   void insertTapePool(const SecurityIdentity &requester, const std::string &name, const uint16_t nbDrives, const uint32_t nbPartialTapes, const std::string &comment);
   
-  void insertStorageClass(const SecurityIdentity &requester, const std::string &name, const uint8_t nbCopies, const std::string &comment);
+  void insertStorageClass(const SecurityIdentity &requester, const std::string &name, const uint16_t nbCopies, const std::string &comment);
   
-  void insertArchiveRoute(const SecurityIdentity &requester, const std::string &storageClassName, const uint8_t copyNb, const std::string &tapePoolName, const std::string &comment);
+  void insertArchiveRoute(const SecurityIdentity &requester, const std::string &storageClassName, const uint16_t copyNb, const std::string &tapePoolName, const std::string &comment);
   
   void insertFile(const SecurityIdentity &requester, const std::string &pathname, const uint16_t mode);
   
@@ -44,7 +44,7 @@ public:
   
   void deleteStorageClass(const SecurityIdentity &requester, const std::string &name);
 
-  void deleteArchiveRoute(const SecurityIdentity &requester, const std::string &storageClassName, const uint8_t copyNb);
+  void deleteArchiveRoute(const SecurityIdentity &requester, const std::string &storageClassName, const uint16_t copyNb);
   
   void deleteFile(const SecurityIdentity &requester, const std::string &pathname);
   
@@ -97,7 +97,7 @@ private:
   
   void checkStorageClassExists(const std::string &name);
   
-  void checkArchiveRouteExists(const std::string &name, const uint8_t copyNb);
+  void checkArchiveRouteExists(const std::string &name, const uint16_t copyNb);
   
   void checkFileExists(const std::string &path, const std::string &name);
